+++ src/test/java/bookstore/controller/ShoppingCartControllerTest.java
@@ -8,14 +8,10 @@ import static org.springframework.test.web.servlet.request.MockMvcRequestBuilder
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
-import bookstore.dto.cartitem.CartItemDto;
 import bookstore.dto.cartitem.CartItemRequestDto;
 import bookstore.dto.shoppingcart.ShoppingCartDto;
-import bookstore.model.ShoppingCart;
-import bookstore.model.User;
+import bookstore.util.ShoppingCartTestUtilClass;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import java.util.ArrayList;
-import java.util.List;
 import org.apache.commons.lang3.builder.EqualsBuilder;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.DisplayName;
@@ -23,8 +19,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.http.MediaType;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.test.context.support.WithUserDetails;
 import org.springframework.test.context.jdbc.Sql;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.MvcResult;
@@ -34,6 +29,7 @@ import org.springframework.web.context.WebApplicationContext;
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 public class ShoppingCartControllerTest {
     protected static MockMvc mockMvc;
+    private final ShoppingCartTestUtilClass testUtil = new ShoppingCartTestUtilClass();
     @Autowired
     private ObjectMapper objectMapper;
 
@@ -48,22 +44,15 @@ public class ShoppingCartControllerTest {
 
     @Test
     @DisplayName("Add cart item")
-    @Sql(scripts = "classpath:database/books/insert-data.sql",
-            executionPhase = Sql.ExecutionPhase.BEFORE_TEST_METHOD)
+    @WithUserDetails(value = "test@gmail.com")
+    @Sql(scripts = "classpath:database/books/insert-data.sql")
     @Sql(scripts =
             "classpath:database/books/cleanup.sql",
             executionPhase = Sql.ExecutionPhase.AFTER_TEST_METHOD)
     void save_ValidRequestDto_Success() throws Exception {
-        User user = createTestUser();
+        CartItemRequestDto requestDto = testUtil.createTestCartItemRequestDto();
 
-        UsernamePasswordAuthenticationToken auth =
-                new UsernamePasswordAuthenticationToken(user, null, List.of());
-
-        SecurityContextHolder.getContext().setAuthentication(auth);
-
-        CartItemRequestDto requestDto = createTestCartItemRequestDto();
-
-        ShoppingCartDto expected = createTestShoppingCartDto();
+        ShoppingCartDto expected = testUtil.createTestShoppingCartDto();
 
         String jsonRequest = objectMapper.writeValueAsString(requestDto);
 
@@ -78,24 +67,17 @@ public class ShoppingCartControllerTest {
                 result.getResponse().getContentAsString(), ShoppingCartDto.class);
         assertNotNull(actual);
         assertNotNull(actual.getId());
-        assertEquals(actual.getId(), expected.getId());
+        assertEquals(actual.getUserId(), expected.getUserId());
     }
 
     @Test
     @DisplayName("Add cart item")
-    @Sql(scripts = "classpath:database/books/insert-data.sql",
-            executionPhase = Sql.ExecutionPhase.BEFORE_TEST_METHOD)
+    @WithUserDetails(value = "test@gmail.com")
+    @Sql(scripts = "classpath:database/books/insert-data.sql")
     @Sql(scripts =
             "classpath:database/books/cleanup.sql",
             executionPhase = Sql.ExecutionPhase.AFTER_TEST_METHOD)
     void save_CartItemDtoIsNull_ReturnsBadRequest() throws Exception {
-        User user = createTestUser();
-
-        UsernamePasswordAuthenticationToken auth =
-                new UsernamePasswordAuthenticationToken(user, null, List.of());
-
-        SecurityContextHolder.getContext().setAuthentication(auth);
-
         CartItemRequestDto requestDto = null;
 
         String jsonRequest = objectMapper.writeValueAsString(requestDto);
@@ -114,20 +96,13 @@ public class ShoppingCartControllerTest {
 
     @Test
     @DisplayName("Find shopping cart")
-    @Sql(scripts = "classpath:database/books/insert-data.sql",
-            executionPhase = Sql.ExecutionPhase.BEFORE_TEST_METHOD)
+    @WithUserDetails(value = "test@gmail.com")
+    @Sql(scripts = "classpath:database/books/insert-data.sql")
     @Sql(scripts =
             "classpath:database/books/cleanup.sql",
             executionPhase = Sql.ExecutionPhase.AFTER_TEST_METHOD)
     void find_ValidAuthentication_Success() throws Exception {
-        User user = createTestUser();
-
-        UsernamePasswordAuthenticationToken auth =
-                new UsernamePasswordAuthenticationToken(user, null, List.of());
-
-        SecurityContextHolder.getContext().setAuthentication(auth);
-
-        ShoppingCartDto expected = createTestShoppingCartDto();
+        ShoppingCartDto expected = testUtil.createTestShoppingCartDto();
 
         MvcResult result = mockMvc.perform(get("/cart")
                 .contentType(MediaType.APPLICATION_JSON))
@@ -139,47 +114,5 @@ public class ShoppingCartControllerTest {
 
         assertNotNull(actual);
         assertTrue(EqualsBuilder.reflectionEquals(expected, actual, "cartItems"));
-    }
-
-    private User createTestUser() {
-        User user = new User();
-        user.setId(1L);
-        user.setEmail("test@gmail.com");
-        user.setPassword("test1pass");
-        user.setFirstName("First");
-        user.setLastName("Last");
-        return user;
-    }
-
-    private CartItemDto createTestCartItemDto() {
-        CartItemDto cartItemDto = new CartItemDto();
-        cartItemDto.setId(1L);
-        cartItemDto.setBookId(1L);
-        cartItemDto.setBookTitle("Witcher");
-        cartItemDto.setQuantity(1);
-        return cartItemDto;
-    }
-
-    private ShoppingCart createTestShoppingCart() {
-        ShoppingCart shoppingCart = new ShoppingCart();
-        shoppingCart.setId(1L);
-        shoppingCart.setUser(createTestUser());
-        return shoppingCart;
-    }
-
-    private CartItemRequestDto createTestCartItemRequestDto() {
-        CartItemRequestDto cartItemRequestDto = new CartItemRequestDto();
-        cartItemRequestDto.setBookId(1L);
-        cartItemRequestDto.setQuantity(1);
-        return cartItemRequestDto;
-    }
-
-    private ShoppingCartDto createTestShoppingCartDto() {
-        ShoppingCartDto shoppingCartDto = new ShoppingCartDto();
-        shoppingCartDto.setId(1L);
-        shoppingCartDto.setUserId(1L);
-        shoppingCartDto.setCartItems(new ArrayList<>());
-        shoppingCartDto.getCartItems().add(createTestCartItemDto());
-        return shoppingCartDto;
     }
 }
