+++ src/test/java/bookstore/service/ShoppingCartServiceTest.java
@@ -5,7 +5,6 @@ import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.mockito.Mockito.verifyNoMoreInteractions;
 import static org.mockito.Mockito.when;
 
-import bookstore.dto.cartitem.CartItemDto;
 import bookstore.dto.cartitem.CartItemRequestDto;
 import bookstore.dto.shoppingcart.ShoppingCartDto;
 import bookstore.exception.EntityNotFoundException;
@@ -13,17 +12,13 @@ import bookstore.mapper.CartItemMapper;
 import bookstore.mapper.ShoppingCartMapper;
 import bookstore.model.Book;
 import bookstore.model.CartItem;
-import bookstore.model.Category;
 import bookstore.model.ShoppingCart;
 import bookstore.model.User;
 import bookstore.repository.BookRepository;
 import bookstore.repository.CartItemRepository;
 import bookstore.repository.ShoppingCartRepository;
-import java.math.BigDecimal;
-import java.util.HashSet;
-import java.util.List;
+import bookstore.util.ShoppingCartTestUtilClass;
 import java.util.Optional;
-import java.util.Set;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -35,6 +30,8 @@ import org.springframework.security.test.context.support.WithMockUser;
 
 @ExtendWith(MockitoExtension.class)
 public class ShoppingCartServiceTest {
+    private final ShoppingCartTestUtilClass testUtil = new ShoppingCartTestUtilClass();
+
     @InjectMocks
     private ShoppingCartServiceImpl shoppingCartService;
 
@@ -62,19 +59,19 @@ public class ShoppingCartServiceTest {
             Verify add cart item to shopping cart
             """)
     public void addCartItem_ValidCartItemRequestDto_ReturnsShoppingCartDto() {
-        User user = createTestUser();
-        Book book = createTestBook();
-        CartItemRequestDto requestDto = createTestCartItemRequestDto();
-        CartItem cartItem = createTestCartItem();
-        CartItem nullCartItem = null;
-        ShoppingCart shoppingCart = createTestShoppingCart();
-        ShoppingCartDto shoppingCartDto = createTestShoppingCartDto();
+        User user = testUtil.createTestUser();
+        Book book = testUtil.createTestBook();
+        CartItemRequestDto requestDto = testUtil.createTestCartItemRequestDto();
+        CartItem cartItem = testUtil.createTestCartItem();
+
+        ShoppingCart shoppingCart = testUtil.createTestShoppingCart();
+        ShoppingCartDto shoppingCartDto = testUtil.createTestShoppingCartDto();
 
         when(authentication.getPrincipal()).thenReturn(user);
 
         when(cartItemRepository.findByShoppingCartIdWhereBookId(
                 user.getId(), requestDto.getBookId()))
-                .thenReturn(nullCartItem);
+                .thenReturn(null);
 
         when(cartItemMapper.toModel(requestDto)).thenReturn(cartItem);
 
@@ -99,11 +96,11 @@ public class ShoppingCartServiceTest {
             Verify add cart item to shopping cart when cart item already exists
             """)
     public void addCartItem_CartItemAlreadyExists_ReturnsShoppingCartDto() {
-        User user = createTestUser();
-        CartItemRequestDto requestDto = createTestCartItemRequestDto();
-        CartItem cartItem = createTestCartItem();
-        ShoppingCart shoppingCart = createTestShoppingCart();
-        ShoppingCartDto shoppingCartDto = createTestShoppingCartDto();
+        User user = testUtil.createTestUser();
+        CartItemRequestDto requestDto = testUtil.createTestCartItemRequestDto();
+        CartItem cartItem = testUtil.createTestCartItem();
+        ShoppingCart shoppingCart = testUtil.createTestShoppingCart();
+        ShoppingCartDto shoppingCartDto = testUtil.createTestShoppingCartDto();
 
         when(authentication.getPrincipal()).thenReturn(user);
 
@@ -132,9 +129,9 @@ public class ShoppingCartServiceTest {
             Verify find cart item in shopping cart
             """)
     public void find_ValidAuthentication_ReturnsShoppingCartDto() {
-        User user = createTestUser();
-        ShoppingCart shoppingCart = createTestShoppingCart();
-        ShoppingCartDto shoppingCartDto = createTestShoppingCartDto();
+        User user = testUtil.createTestUser();
+        ShoppingCart shoppingCart = testUtil.createTestShoppingCart();
+        ShoppingCartDto shoppingCartDto = testUtil.createTestShoppingCartDto();
 
         when(authentication.getPrincipal()).thenReturn(user);
 
@@ -156,12 +153,12 @@ public class ShoppingCartServiceTest {
     public void updateCartItemById_ValidCartItem_ReturnsShoppingCartDto() {
         Long id = 1L;
         int quantity = 2;
-        User user = createTestUser();
-        CartItem cartItem = createTestCartItem();
-        CartItem updatedCartItem = createTestCartItem();
+        User user = testUtil.createTestUser();
+        CartItem cartItem = testUtil.createTestCartItem();
+        CartItem updatedCartItem = testUtil.createTestCartItem();
         updatedCartItem.setQuantity(quantity);
-        ShoppingCart shoppingCart = createTestShoppingCart();
-        ShoppingCartDto shoppingCartDto = createTestShoppingCartDto();
+        ShoppingCart shoppingCart = testUtil.createTestShoppingCart();
+        ShoppingCartDto shoppingCartDto = testUtil.createTestShoppingCartDto();
 
         when(authentication.getPrincipal()).thenReturn(user);
 
@@ -187,7 +184,7 @@ public class ShoppingCartServiceTest {
             Verify update cart item to shopping cart invalid cart item id
             """)
     public void updateCartItemById_InvalidCartItemId_ReturnsException() {
-        User user = createTestUser();
+        User user = testUtil.createTestUser();
         Long wrongId = 2L;
         int quantity = 2;
 
@@ -198,66 +195,5 @@ public class ShoppingCartServiceTest {
 
         assertThrows(EntityNotFoundException.class,
                 () -> shoppingCartService.updateCartItemById(wrongId, quantity, authentication));
-    }
-
-    private User createTestUser() {
-        User user = new User();
-        user.setId(1L);
-        user.setEmail("test@gmail.com");
-        user.setPassword("test1pass");
-        user.setFirstName("First");
-        user.setLastName("Last");
-        return user;
-    }
-
-    private Book createTestBook() {
-        Book book = new Book();
-        book.setId(1L);
-        book.setTitle("Witcher");
-        book.setAuthor("Sapkovsky");
-        book.setIsbn("1981112314470");
-        book.setPrice(BigDecimal.valueOf(30.75));
-        Set<Category> categories = new HashSet<>();
-        return book;
-    }
-
-    private CartItem createTestCartItem() {
-        CartItem cartItem = new CartItem();
-        cartItem.setId(1L);
-        cartItem.setShoppingCart(new ShoppingCart());
-        cartItem.setBook(createTestBook());
-        cartItem.setQuantity(1);
-        return cartItem;
-    }
-
-    private CartItemRequestDto createTestCartItemRequestDto() {
-        CartItemRequestDto cartItemRequestDto = new CartItemRequestDto();
-        cartItemRequestDto.setBookId(1L);
-        cartItemRequestDto.setQuantity(1);
-        return cartItemRequestDto;
-    }
-
-    private CartItemDto createTestCartItemDto() {
-        CartItemDto cartItemDto = new CartItemDto();
-        cartItemDto.setId(1L);
-        cartItemDto.setBookId(1L);
-        cartItemDto.setBookTitle("Witcher");
-        cartItemDto.setQuantity(1);
-        return cartItemDto;
-    }
-
-    private ShoppingCart createTestShoppingCart() {
-        ShoppingCart shoppingCart = new ShoppingCart();
-        shoppingCart.setId(1L);
-        shoppingCart.setUser(createTestUser());
-        return shoppingCart;
-    }
-
-    private ShoppingCartDto createTestShoppingCartDto() {
-        ShoppingCartDto shoppingCartDto = new ShoppingCartDto();
-        shoppingCartDto.setId(1L);
-        shoppingCartDto.setUserId(1L);
-        shoppingCartDto.setCartItems(List.of(createTestCartItemDto()));
-        return shoppingCartDto;
     }
 }
